move remaining impala ops to base_sql
